LayoutTests/crypto/workers/subtle/sha384-digest.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
    <script>
        worker = startWorker('resources/sha384-digest.js');
    </script>
    <script src="../../../resources/js-test-post.js"></script>
</body>
</html>
